requirements.txt
python-telegram-bot[ext]>=20.0
python-dotenv
google-api-python-client
google-auth-oauthlib
